requirements/base.txt
alembic==1.5.4
aniso8601==8.1.1
appdirs==1.4.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cov-core==1.15.0
coverage==5.4
Flask==1.1.2
Flask-HTTPAuth==4.2.0
Flask-JWT-Extended==3.25.0
flask-marshmallow==0.14.0
Flask-Migrate==2.6.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
httpie==2.4.0
idna==2.10
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.2
mypy-extensions==0.4.3
nose2==0.10.0
passlib==1.7.4
pathspec==0.8.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
Pygments==2.7.4
PyJWT==1.7.1
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
regex==2020.11.13
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
